moved ReadFromLink to the shared/gcs
