+++ shared/gcs/gcs_test.go
